feat(Engine): creation of the game engine
Engine.hpp/cpp : Game Engine class

RenderGame.hpp: called control_system when no event is currently called
set a framerate limit
deleted comments

System.hpp : created an hpp file, because having everything in a cpp was ass

Position.hpp: Changed values from int to uint_16_t

Velocity.hpp: Changed values from int to uint_16_t

main.cpp : new tests, but we don't care about the main

+++ client/include/components/Position.hpp
@@ -9,7 +9,7 @@
 
 struct Position {
     Position() = default;
-    void build_component(const int &x, const int &y) {
+    void build_component(const uint16_t &x, const uint16_t &y) {
         _x = x;
         _y = y;
     }
@@ -17,6 +17,6 @@ struct Position {
         output << "x = " << D._x << " y = " <<  D._y << std::endl;
         return output;
     }
-    int _x;
-    int _y;
+    uint16_t _x;
+    uint16_t _y;
 };
